djangogram/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
behave==1.2.6
behave-django==1.4.0
Django==3.1.7
django-tables2==2.3.4
djangorestframework==3.12.4
parse==1.19.0
parse-type==0.5.2
Pillow==8.1.2
pytz==2021.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
